added contact form

+++ services/views.py
@@ -1,6 +1,6 @@
 from django.shortcuts import render, get_object_or_404, redirect, reverse
 from django.contrib import messages
-
+from django.contrib.auth.decorators import login_required
 from .models import Service, Category
 
 from .forms import ServiceForm
@@ -38,8 +38,13 @@ def service_detail(request, service_id):
     return render(request, 'services/service_details.html', context)
 
 
+@login_required
 def add_service(request):
     """ Add a Service to the store"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can use this')
+        return redirect(reverse('home'))
+
     if request.method == "POST":
         form = ServiceForm(request.POST, request.FILES)
         if form.is_valid():
@@ -61,8 +66,13 @@ def add_service(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_service(request, service_id):
     """Edit a Service in the store"""
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can use this')
+        return redirect(reverse('home'))
 
     service = get_object_or_404(Service, pk=service_id)
 
@@ -87,8 +97,15 @@ def edit_service(request, service_id):
 
     return render(request, template, context)
 
+
+@login_required
 def delete_service(request, service_id):
     """ Delete a Service from the store """
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can use this')
+        return redirect(reverse('home'))
+
     service = get_object_or_404(Service, pk=service_id)
     service.delete()
     messages.success(request, 'Succesfully deleted'
